feat: unified hidden/internal entity system
